split out system and mock scheduler
