+++ src/socket/web/server.ts
@@ -9,7 +9,7 @@ import {
     MuSocketServerSpec,
 } from '../../socket';
 import { MuScheduler } from '../../scheduler/scheduler';
-import { MuSystemScheduler } from '../../scheduler';
+import { MuSystemScheduler } from '../../scheduler/system';
 
 export interface UWSSocketInterface {
     onmessage:(message:{ data:Uint8Array|string }) => void;
