+++ src/socket/worker/server.ts
@@ -11,7 +11,7 @@ import {
     MuCloseHandler,
 } from '../../socket';
 import { MuScheduler } from '../../scheduler/scheduler';
-import { MuSystemScheduler } from '../../scheduler';
+import { MuSystemScheduler } from '../../scheduler/system';
 
 function noop () { }
